+++ NEWS.md
@@ -1,2 +1,3 @@
+-   climfor 0.1.0.9003: tmf deforest & kendall's tau for chirps
 -   climfor 0.1.0.9002: quick exploration with dl script def
 -   climfor 0.1.0.9001: init